Added home button, redirect home page based on loggin, surveys each have a line, deleted extra text.
The home route '/' directed to the users index page previously so that users could sign in. We moved the sign in page to its own erb, had the route render that instead and also made it check the status of loggin so that if a user is logged in they go straight to the surveys index page.